Implemented WAT version of query from figure 2.3 of WAM tutorial.
Implemented WAM registers as globals. Global $Xn is the n-th register. There is a pair of functions for getting and setting each global register and these functions are listed in the Table so that call_indirect can be used to access a register parametrically.
Another option of registers is to use Memory slots for them, e.g. slots 0..3 for (word) register 1, 4..7 for register 2, etc. Then i32.load and i32.store can be used to access these registers.