Implemented M2 and L2 of tutorial in language2.wat and test_wam2.html.
The language2.wat adds an 'instruction word interpreter' to evaluate 'code' arrays held in the host Javascript accessible through the 'getCode' function.
The initial code array must be a query and it must end with a call instruction for a program.
The program(s) must have their code selectable through the 'setCode' functions (called by language2.wat).
The test_wam2.html file implements the program of the figure 3.1 of the wam tutorial and solutions for exercise 3.1.