Changed inference counting to use incomplete inferences for execute inst.
The proceed instruction adds 1 plus the number of outstanding incomplete inferences in the most recent choicepoint frame or the 'base' (pre choicepoints) and resets to 0 the count of incomplete inferences. Each execute instruction adds 1 to the incomplete inferences count in the choicepoint frame or the 'base' (pre choicepoints). Each backtrack resets the incomplete inferences count (of the current choicepoint frame, if any) to 0.